requirements.txt
Django>=5.0,<5.1
django-tailwind[reload]==3.6.0
django-allauth>0.61, <1.0
django-ninja>=1.1, <2.0